implement limited live update functionality for posts
There is still no live update for all posts (i.e. subscription),
but if a user clicks on a post
and then returns, this post will be refeteched.
This should allow, that the user (if he has the right for it) can make
changes to the post, and these changes will be diretly visible
when returning